fixing message and others

+++ src/main/java/com/enterpriseintellijence/enterpriseintellijence/data/services/MessageServiceImp.java
@@ -1,16 +1,25 @@
 package com.enterpriseintellijence.enterpriseintellijence.data.services;
 
 import com.enterpriseintellijence.enterpriseintellijence.data.entities.Message;
+import com.enterpriseintellijence.enterpriseintellijence.data.entities.Product;
 import com.enterpriseintellijence.enterpriseintellijence.data.entities.User;
 import com.enterpriseintellijence.enterpriseintellijence.data.repository.MessageRepository;
+import com.enterpriseintellijence.enterpriseintellijence.data.repository.ProductRepository;
+import com.enterpriseintellijence.enterpriseintellijence.data.repository.UserRepository;
 import com.enterpriseintellijence.enterpriseintellijence.dto.MessageDTO;
 import com.enterpriseintellijence.enterpriseintellijence.dto.UserDTO;
+import com.enterpriseintellijence.enterpriseintellijence.dto.enums.MessageStatus;
+import com.enterpriseintellijence.enterpriseintellijence.dto.enums.UserRole;
 import com.enterpriseintellijence.enterpriseintellijence.exception.IdMismatchException;
 
+import com.enterpriseintellijence.enterpriseintellijence.security.JwtContextUtils;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
+
+import java.time.Clock;
+import java.time.LocalDateTime;
 
 @Service
 @RequiredArgsConstructor
@@ -18,27 +27,43 @@ public class MessageServiceImp implements MessageService{
 
 
     private final MessageRepository messageRepository;
+    private final JwtContextUtils jwtContextUtils;
+    private final ProductRepository productRepository;
 
+    private final UserRepository userRepository;
     private final UserService userService;
+    private final Clock clock;
 
     private final ModelMapper modelMapper;
 
     @Override
     public MessageDTO createMessage(MessageDTO messageDTO) {
         Message message = mapToEntity(messageDTO);
+        message.setId(null);
+        User loggedUser = jwtContextUtils.getUserLoggedFromContext();
+        // TODO: 23/05/2023 l'utente che invia lo prendo dal jwtcontext...dovrei confrontarlo con l'id del messaggioDTO.getSendUser?
+        User receivedUser = userRepository.findById(messageDTO.getReceivedUser().getId()).orElseThrow(EntityNotFoundException::new);
 
-        UserDTO userDTO = userService.findUserFromContext()
-                .orElseThrow(EntityNotFoundException::new);
+        message.setSendUser(loggedUser);
+        message.setReceivedUser(receivedUser);
+        message.setMessageDate(LocalDateTime.now(clock));
 
-        message.setSendUser(modelMapper.map(userDTO, User.class));
+        Product product;
+        if(messageDTO.getProduct()!=null) {
+            product = productRepository.findById(messageDTO.getProduct().getId()).orElseThrow(EntityNotFoundException::new);
+            message.setProduct(product);
+        }
+        message.setMessageStatus(MessageStatus.UNREAD);
 
         message = messageRepository.save(message);
+        System.out.println("a chi?");
 
         return mapToDTO(message);
     }
 
     @Override
     public MessageDTO replaceMessage(String id, MessageDTO messageDTO) throws IllegalAccessException {
+        // TODO: 23/05/2023 ha senso che ci sia?
         throwOnIdMismatch(id, messageDTO);
 
         Message oldMessage = messageRepository.findById(id).orElseThrow(EntityNotFoundException::new);
@@ -64,6 +89,7 @@ public class MessageServiceImp implements MessageService{
 
     @Override
     public MessageDTO updateMessage(String id, MessageDTO patch) {
+        // TODO: 23/05/2023 ha senso che ci sia?
         MessageDTO message = mapToDTO(messageRepository.findById(id).orElseThrow(EntityNotFoundException::new));
 
         if (patch.getContext() != null) {
@@ -83,7 +109,18 @@ public class MessageServiceImp implements MessageService{
     }
 
     @Override
-    public void deleteMessage(String id) {
+    public void deleteMessage(String id) throws IllegalAccessException {
+        Message message = messageRepository.findById(id).orElseThrow(EntityNotFoundException::new );
+        User loggedUser = jwtContextUtils.getUserLoggedFromContext();
+
+        if(!loggedUser.getRole().equals(UserRole.USER) && (loggedUser.getId().equals(message.getSendUser().getId()) || loggedUser.getId().equals(message.getReceivedUser().getId())))
+            throw new IllegalAccessException("Cannot delete others message");
+
+        if(message.getMessageStatus().equals(MessageStatus.READ) && message.getOffer()!=null)
+            throw new IllegalAccessException("Cannot delete message with offers read");
+
+        //if(message.ge)
+
         messageRepository.findById(id).orElseThrow(EntityNotFoundException::new);
         messageRepository.deleteById(id);
     }
